+++ src/widgets/styles/qwindowsstyle.cpp
@@ -856,7 +856,7 @@ void QWindowsStyle::drawPrimitive(PrimitiveElement pe, const QStyleOption *opt, 
             QPointF  points[6];
             qreal scaleh = opt->rect.width() / 12.0;
             qreal scalev = opt->rect.height() / 12.0;
-            points[0] = { opt->rect.x() + 3.5 * scaleh, opt->rect.y() + 5.5 * scalev };
+            points[0] = { opt->rect.x() + qreal(3.5) * scaleh, opt->rect.y() + qreal(5.5) * scalev };
             points[1] = { points[0].x(),                points[0].y() + 2 * scalev };
             points[2] = { points[1].x() + 2 * scaleh,   points[1].y() + 2 * scalev };
             points[3] = { points[2].x() + 4 * scaleh,   points[2].y() - 4 * scalev };
